Automatically backed up by Learn

+++ db/01_create_users_table.rb
@@ -4,7 +4,7 @@ class CreateUsersTable < ActiveRecord::Migration
     create_table :users do |t|
       t.string :username 
       t.string :password
-      t.float :balance 
+      t.integer :balance 
   end
